@@ -1,8 +1,8 @@
 ﻿using Dapper;
 using DataAccess.Models;
+using Microsoft.Data.SqlClient;
 using System.Collections.Generic;
 using System.Data;
-using System.Data.SqlClient;
 
 namespace DataAccess.Repositories
 {
